[Test] Add basic matched stop for beta eagle (#11833)
